feat(wifi): use uint64 for WiFi TCP byte counters (long session support)
Switch wifiTcpBytesSent and wifiTcpBytesConfirmed from uint32 to uint64
to support multi-day streaming sessions without rollover. Already
protected by critical sections (required for 64-bit on PIC32MZ).
Consistent with TotalSamplesStreamed/TotalBytesStreamed which are also
uint64.

uint32 would overflow at ~73 hours at 16 KB/s WiFi throughput.
uint64 is safe for billions of years.

Entire-Checkpoint: 2872097ec67b

+++ firmware/src/services/SCPI/SCPIInterface.c
@@ -1642,8 +1642,8 @@ scpi_result_t SCPI_GetStreamStats(scpi_t * context) {
     {
         wifi_tcp_server_context_t* pTcp = wifi_manager_GetTcpServerContext();
         if (pTcp != NULL) {
-            scpi_printf(context, "WifiTcpBytesSent=%u\r\n", (unsigned)pTcp->client.wifiTcpBytesSent);
-            scpi_printf(context, "WifiTcpBytesConfirmed=%u\r\n", (unsigned)pTcp->client.wifiTcpBytesConfirmed);
+            scpi_printf(context, "WifiTcpBytesSent=%llu\r\n", (unsigned long long)pTcp->client.wifiTcpBytesSent);
+            scpi_printf(context, "WifiTcpBytesConfirmed=%llu\r\n", (unsigned long long)pTcp->client.wifiTcpBytesConfirmed);
             scpi_printf(context, "WifiTcpSendErrors=%u\r\n", (unsigned)pTcp->client.wifiTcpSendErrors);
         }
     }
